backend/pyproject.toml
[project]
name = "budget-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "python-multipart>=0.0.9",
  "passlib[bcrypt]>=1.7",
  "pyjwt[crypto]>=2.9"
]
